Add tests for long autoincrement columns and blobs, fix dialects.

+++ squash-core/test/org/jetbrains/squash/tests/data/CitiesSchema.kt
@@ -25,7 +25,7 @@ object CitizenDataLink : TableDefinition() {
 }
 
 object CitizenData : TableDefinition() {
-    val id = integer("id").autoIncrement().primaryKey()
+    val id = long("id").autoIncrement().primaryKey()
     val comment = varchar("comment", 30)
     val value = enumeration<DataKind>("value")
     val image = blob("image").nullable()
